+++ language/src/main/java/i4gl/runtime/types/compound/CharType.java
@@ -24,7 +24,7 @@ public class CharType extends ArrayType {
 
     @Override
     public Object getDefaultValue() {
-        return new Char(size);
+        return new Char(this);
     }
 
     @Override